update transmittable report
