+++ script/crm_transmittable_report.rb
@@ -11,7 +11,8 @@ if ARGV[0].blank?
 end
 
 limit = ARGV[0].to_i
-result = Operations::CRMTransmittableProcessor.new.call(limit)
+date = ARGV[1]&.to_date || Date.today
+result = Operations::SugarCRM::GenerateReport.new.call({limit: limit, date: date})
 
 if result.success?
   puts result.success
